Move all double-scene hooks to container file
