+++ src/components/double-scene/double-scene-component.jsx
@@ -1,82 +1,23 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
-import { loadModules } from 'esri-loader';
 import Spinner from 'components/spinner';
 import styles from 'styles/themes/scene-theme.module.scss';
 
-const DoubleSceneComponent = ({ sceneId, children, loaderOptions, sceneSettings, onMapLoad = null, onViewLoad = null, style, spinner = true, interactionsDisabled = false, sceneMode, countryExtent }) => {
-  const [map, setMap] = useState(null);
-  const [viewGlobal, setViewGlobal] = useState(null);
-  const [viewLocal, setViewLocal] = useState(null);
-  const [loadState, setLoadState] = useState('loading');
-  const [spatialReference, setSpatialReference] = useState(null);
-
-  useEffect(() => {
-    loadModules(["esri/WebScene"], loaderOptions)
-      .then(([WebScene]) => {
-        const _map = new WebScene({
-          portalItem: {
-            id: sceneId
-          }
-        });
-        _map.load().then(map => { 
-          setMap(map);
-          onMapLoad && onMapLoad(map);
-        })
-      })
-      .catch(err => {
-        console.error(err);
-      });
-  }, [])
-
-  useEffect(() => {
-    if (map) {
-      loadModules(["esri/views/SceneView", "esri/geometry/SpatialReference"], loaderOptions)
-        .then(([SceneView, SpatialReference]) => {
-
-          const _spatialReference = new SpatialReference();
-          _spatialReference.wkid = 102100;
-          setSpatialReference(_spatialReference);
-
-          const _viewGlobal = new SceneView({
-            map: map,
-            container: `scene-global-container-${sceneId}`,
-            ...sceneSettings,
-            spatialReference
-          });
-          setViewGlobal(_viewGlobal);
-
-          const _viewLocal = new SceneView({
-            map: map,
-            container: `scene-local-container-${sceneId}`,
-            ...sceneSettings,
-            viewingMode: 'local',
-            spatialReference,
-          });
-          setViewLocal(_viewLocal);
-        })
-        .catch(err => {
-          console.error(err);
-        });
-    }
-  },[map]);
-
-  useEffect(() => {
-    if(viewLocal && spatialReference && countryExtent) {
-      const expandedCountryExtent = countryExtent.clone().expand(1.5); //add paddings around country borders
-      viewLocal.clippingArea = expandedCountryExtent;
-      viewLocal.extent = expandedCountryExtent;
-      viewLocal.goTo({ target: expandedCountryExtent, tilt: 40, }, { animate: false });
-    };
-  },[countryExtent]);
+const DoubleSceneComponent = props => {
+  const {
+    map,
+    viewGlobal,
+    viewLocal,
+    spatialReference,
+    spinner,
+    loadState,
+    sceneId,
+    children,
+    interactionsDisabled,
+    style,
+    sceneMode
+  } = props;
   
-  useEffect(() => {
-    if (map && viewGlobal && viewLocal) {
-      setLoadState('loaded');
-      onViewLoad && onViewLoad(map, viewGlobal)
-    }
-  }, [map, viewGlobal, viewLocal]);
-
   return (
     <>
       {loadState === 'loading' && <Spinner spinnerWithOverlay initialLoading display={spinner}/>}
